requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
Django
django-ckeditor-5
django-jazzmin
django-js-asset
django-rest-swagger
djangorestframework
drf-yasg
idna
inflection
itypes
Jinja2
MarkupSafe
openapi-codec
packaging
pillow
python-dotenv
pytz
PyYAML
requests
simplejson
sqlparse
uritemplate
urllib3
whitenoise
